+++ src/biz.dfch.CS.Appclusive.UI/Controllers/JobsController.cs
@@ -34,7 +34,7 @@ namespace biz.dfch.CS.Appclusive.UI.Controllers
         }
 
         // GET: Jobs/Details/5
-        public ActionResult Details(int id)
+        public ActionResult Details(long id)
         {
             try
             {
