+++ src/biz.dfch.CS.Appclusive.UI/Controllers/NodesController.cs
@@ -34,7 +34,7 @@ namespace biz.dfch.CS.Appclusive.UI.Controllers
         }
 
         // GET: Nodes/Details/5
-        public ActionResult Details(int id)
+        public ActionResult Details(long id)
         {
             var item = CoreRepository.Nodes.Expand("Children").Where(c => c.Id == id).FirstOrDefault();
             return View(AutoMapper.Mapper.Map<Models.Core.Node>(item));
